fix some return code

+++ tests/scripts/fusionforge-build-and-test-source.sh
@@ -69,12 +69,9 @@ else
 	scp ../build/packages/fusionforge-*.tar.bz2 root@$HOST:
 fi
 ssh root@$HOST 'tar jxf fusionforge-*.tar.bz2'
-if [ -e "/tmp/timedhosts.txt" ]
-then
-	scp -p /tmp/timedhosts.txt root@$HOST:/var/cache/yum/timedhosts.txt
-fi
+[ ! -e "/tmp/timedhosts.txt" ] || scp -p /tmp/timedhosts.txt root@$HOST:/var/cache/yum/timedhosts.txt
 ssh root@$HOST 'cd fusionforge-*; FFORGE_RPM_REPO=http://buildbot.fusionforge.org/job/fusionforge-trunk-build-and-test-rpm/ws/build/packages/ FFORGE_DB=fforge FFORGE_USER=gforge FFORGE_ADMIN_USER=ffadmin FFORGE_ADMIN_PASSWORD=ffadmin ./install.sh centos52.local'
-scp -p root@$HOST:/var/cache/yum/timedhosts.txt /tmp/timedhosts.txt
+scp -p root@$HOST:/var/cache/yum/timedhosts.txt /tmp/timedhosts.txt || true
 ssh root@$HOST "(echo [core];echo use_ssl=no) > /etc/gforge/config.ini.d/zzz-builbot.ini"
 ssh root@$HOST "cd /root/tests/func; CONFIGURED=true CONFIG_PHP=config.php.buildbot DB_NAME=$DB_NAME php db_reload.php"
 ssh root@$HOST "su - postgres -c \"pg_dump -Fc $DB_NAME\" > /root/dump"
